+++ CHANGELOG.md
@@ -59,6 +59,7 @@
 - Enh #86: Add `HasAutocapitalize` and `HasAutocorrect` traits and `autocapitalize()`, `autocorrect()` methods to manage `autocapitalize` and `autocorrect` attributes for HTML elements (@terabytesoftw)
 - Bug #87: Update test method names for clarity by removing redundant `Value` suffix (@terabytesoftw)
 - Enh #88: Add `CanBeSelected`, `HasLabel` traits and `selected()`, `label()` methods to manage `selected` and `label` attributes for HTML elements (@terabytesoftw)
+- Enh #89: Add `HasFor` trait and `for()` method to manage `for` attribute for HTML elements (@terabytesoftw)
 
 ## 0.5.2 January 29, 2026